some pre-emptive safety features implemented for vision testing
added current limits to all of the drive motors in the event I do something incredibly stupid and lock the robot up. This avoids killing the NEOs. Also re-added the button for vision tracking to driver's right trigger, we'll start with getting the vision we had working earlier in the year back to functional